product-api/requirements.txt
Flask==1.1.1
flask-restful==0.3.5
mysql-connector-python==8.0.18